Use focus.default_sort when appropriate
Current code doesn't expose `focus.default_sort` at all, and calling
code uses `focus.sorts.default` which degenerates to the last entry
in the list of sorts.

This PR exposes focus.default_sort (which itself defaults to
focus.sorts.default if the default_sort is undefined in the configuration
file) and then uses it instead of `sorts.default` in calling code.